// ts/wompo/types.ts
export interface RenderHtml {
  parts: TemplateStringsArray;
  values: any[];
  key?: string;
  _$wompoHtml: true;
  _$wompoSvg?: boolean;
  _$portal?: HTMLElement;
}

/** The props of any component. */
export interface WompoProps {
  children?: WompoChildren;
  styles?: { [key: string]: string };
  ['wc-perf']?: boolean;
  wcPerf?: boolean;
  style?: string | Partial<CSSStyleDeclaration> | object;
  ref?: RefHook<any>;
  id?: string;
  class?: string;
  [event: `on${string}`]: (ev: Event) => void;
}

/** The options that a component can have when instantiating. */
export interface WompoComponentOptions {
  name?: string;
  shadow?: boolean;
  cssModule?: boolean;
  /** Default hydration trigger. When set, every instance of this component is treated as an
   * island and hydrated according to this mode unless overridden by a `client:*` attribute on
   * the call site. */
  island?: 'load' | 'idle' | 'visible';
}

/** Wompo component function with extra metadata fields populated by `defineWompo`. */
export interface WompoComponent<Props extends WompoProps = WompoProps> {
  (props: Props): RenderHtml;
  css?: string;
  componentName?: string;
  _$wompoF?: true;
  class?: WompoElementClass<Props>;
  options?: {
    generatedCSS: string;
    styles: { [key: string]: string };
    shadow: boolean;
    island?: 'load' | 'idle' | 'visible';
  };
}

/** A Wompo component instance (HTMLElement + internal fields). */
export type WompoElement<Props extends WompoProps = WompoProps, E = {}> = HTMLElement &
  E & {
    props: Props;
    hooks: Hook[];
    _$initialProps: WompoProps;
    _$portals: any[];
    _$measurePerf: boolean;
    _$usesContext: boolean;
    _$hasBeenMoved: boolean;
    _$effects: EffectHook[];
    _$asyncCalls: AsyncHook<any>[];
    _$suspendedAsyncCalls: AsyncHook<any>[];
    _$layoutEffects: EffectHook[];
    requestRender: () => void;
    onDisconnected: () => void;
    updateProp: (prop: string, newValue: any) => void;
    _$wompo: true;
    _$updating: boolean;
  };

export type Hook =
  | StateHook<any>
  | EffectHook
  | RefHook<any>
  | CallbackHook<any>
  | IdHook
  | MemoHook<any>
  | ReducerHook<any>
  | AsyncHook<any>
  | ContextHook;

export type StateHook<S> = [S, (newValue: S | ((oldValue: S) => S)) => void];

export interface EffectHook {
  dependencies: any;
  callback: VoidFunction | (() => VoidFunction);
  cleanupFunction: VoidFunction | void;
}

export interface RefHook<V> {
  current: V;
  __wcRef: true;
}

export interface CallbackHook<C> {
  dependencies?: any[];
  value: C;
}

export type IdHook = string;

export interface MemoHook<T = any> {
  dependencies: any[];
  value: T;
}

export interface ReducerAction {
  [key: string]: any;
}

export type ReducerHook<State> = [State, (action: ReducerAction) => void];

export interface AsyncHook<S> {
  asyncCallback: () => Promise<S>;
  dependencies: any[];
  value: S;
  activateSuspense: boolean;
}

export interface ContextProviderProps extends WompoProps {
  value: any;
}
export interface ContextProviderExposed {
  subscribers: RefHook<Set<WompoElement>>;
}
export type ContextProviderElement = WompoElement<ContextProviderProps, ContextProviderExposed>;

export interface ContextHook {
  node: ContextProviderElement;
}

/** The type of the class generated by the wompo() function. */
export interface WompoElementClass<Props extends WompoProps, E = {}> {
  new (): WompoElement<Props, E>;
  _$cachedTemplate: any;
  _$getOrCreateTemplate(parts: TemplateStringsArray): any;
}